Move CustomActionOperationMixin to a dedicated mixin file
